requirements.txt
# Core application dependencies
streamlit>=1.28.0
openai-whisper>=20231117
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
python-docx>=0.8.11
pydub>=0.25.1
numpy>=1.24.0

# PyTorch with CUDA support for GPU acceleration (RTX 4090 optimized)
# Use CUDA 12.1 compatible versions for optimal RTX 4090 performance
torch>=2.1.0,<2.6.0
torchaudio>=2.1.0,<2.6.0

# Additional GPU acceleration dependencies
nvidia-ml-py3>=7.352.0  # NVIDIA Management Library for GPU monitoring
# Note: Triton kernels not available on Windows - warning can be ignored
